+++ src/addons/character_dna/__init__.py
@@ -17,7 +17,7 @@ logger = logging.getLogger(constants.ToolInfo.NAME)
 bl_info = {
     "name": "Character DNA",
     "author": "Poly Hammer",
-    "version": (0, 12, 12),
+    "version": (0, 12, 13),
     "blender": (4, 5, 0),
     "location": "File > Import > MetaHuman DNA",
     "description": (